Redirect to error view when accessing unknown child
The user can login to the service using other services like Google or
Facebook. Once they have logged in using a service, they must always
use the same one if they want to access the "kukkuu" account that is
created then.

Sometimes users login with a service they did not use during their
initial sign up process. In this scenario, they are not able to access
their original data.

A usecase that's especially problematic is when the user receives a
link into their mail they can use to sign up their child into an event.
If they login by using the wrong service, they end up in a view that
doesn't work and gives little in the way of clues as to why.

After this commit, the user would be redirected into the wrong login
method page, which guides them to sign in again.