Use replace when redirecting user according to state
If push is used instead of replace, the callback page is retained
within the history. The user can then access it by using the back
button, which will lead to an error.

These errors are problematic, because they flood Sentry and may be
hiding other usecases which lead to the same error. The user experience
when backing into the callback page is also poor in general and may
mislead the user.