Fix empty child list on mobile
Previously the logged in user's profile was synced into the redux state
when the details for their user menu were fetched. On mobile, this menu
is not rendered by default, so their details were not ever synced and
they ended up seeing an empty list of children instead.

It's a bit odd that the user details are stored in two places: (1) in
Apollo's state and in (2) Kukkuu's redux state. I think that ideally
this duplication would not exist. But because I don't understand enough
of the context of the application yet, I didn't remove this
duplication entirely.

As a fix I created the useProfile hook. It fetches the profile and
stores it into the redux state--roughly the logic that was contained
within the user menu before. After this change, the profile is loaded
whenever it is accessed so there's no chance that a view will request
the profile from the state without a requested with the server never
beginning. Further, this structure will make the consumers of profile
data more agnostic about its source which will make it easier to
change responsibilities between the states later on.